+++ src/tools/teledisk/lzhtest.cpp
@@ -16,7 +16,7 @@ unsigned int TestEncode(unsigned char *in, unsigned int inlen,
     int count = 0;
     int in_block_size = 1;
     int out_block_size = 1;
-    for(int i = 0; i < inlen;) {
+    for(unsigned int i = 0; i < inlen;) {
         unsigned int _inlen = std::min((int)(inlen-i), in_block_size);
         unsigned int _outlen = std::min((int)(outlen-count), out_block_size);
         lzh::Encode(in+i, _inlen, out+count, _outlen, context);
@@ -44,12 +44,12 @@ unsigned int TestDecode(unsigned char *in, unsigned int inlen, unsigned char *ou
 
     lzh::DecodeContext *context = lzh::BeginDecode();
 
-    int count = 0;
-    int in_block_size = 1;
-    int out_block_size = 1;
-    for(int i = 0; i < inlen;) {
-        unsigned int _inlen = std::min((int)(inlen-i), in_block_size);
-        unsigned int _outlen = std::min((int)(outlen-count), out_block_size);
+    unsigned int count = 0;
+    unsigned int in_block_size = 1;
+    unsigned int out_block_size = 1;
+    for(unsigned int i = 0; i < inlen;) {
+        unsigned int _inlen = std::min(inlen-i, in_block_size);
+        unsigned int _outlen = std::min(outlen-count, out_block_size);
         lzh::Decode(in+i, _inlen, out+count, _outlen, context);
         i += _inlen;
         count += _outlen;
@@ -57,7 +57,7 @@ unsigned int TestDecode(unsigned char *in, unsigned int inlen, unsigned char *ou
 
     while(count < outlen) {
         unsigned int _inlen = 0;
-        unsigned int _outlen = std::min((int)(outlen-count), out_block_size);
+        unsigned int _outlen = std::min(outlen-count, out_block_size);
         lzh::Decode(nullptr, _inlen, out+count, _outlen, context);
         count += _outlen;
     }
@@ -91,7 +91,7 @@ int main(int argc, char**argv) {
     fprintf(stderr,"Compressed to %zu bytes\n", compr.size());
     TestDecode((unsigned char*)compr.data(), compr.size(), (unsigned char*)uncompr.data(), uncompr.size());
     if(memcmp(uncompr.data(), src.data(), src.size()) != 0) {
-        for(int i = 0; i < src.size(); i++) {
+        for(unsigned int i = 0; i < src.size(); i++) {
             if(uncompr[i] != src[i]) {
                 fprintf(stderr, "Original and uncompressed mismatch at position %d\n", i);
                 break;
